+++ backend/src/messages/messages.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { MessagesService } from './messages.service';
+import { MessagesGateway } from './messages.gateway';
+import { MessagesController } from './messages.controller';
+
+@Module({
+  controllers: [MessagesController],
+  providers: [MessagesGateway, MessagesService],
+})
+export class MessagesModule {}